Restructured the backend codebase
