+++ Main-Folder/Mboalab-Project/Backend/DB/connect.js
@@ -1,0 +1,12 @@
+const mongoose = require("mongoose")
+
+const connectDB = (url) => {
+    return mongoose.connect(url, { 
+        useNewUrlParser: true,
+        useUnifiedTopology: true
+    }).then(() => {
+        console.log('Connected!!!');
+      });
+}
+
+module.exports = connectDB
