+++ Main-Folder/Mboalab-Project/Backend/README.md
@@ -1,0 +1,27 @@
+# Getting Started with Node.js App
+
+## Instructions to use the app
+
+To add all dependencies:
+
+### `npm install`
+
+## Available Scripts
+
+In the project directory, you can run the server using:
+
+### `npm start`
+
+Runs the app in the development mode.\
+Open [http://localhost:5000](http://localhost:5000) to view it in your browser.
+
+# Endpoints
+The following endpoints are available on this server 
+
+ `/contact/` : to add a new user message.
+
+ `/service/` : to add a data checkup survey.
+
+ `/comments` : to add a new comment .
+
+ `/allcomments` : to retrieve all comments.
