+++ Main-Folder/Mboalab-Project/Backend/config.env
@@ -1,0 +1,2 @@
+MONGO_URI = mongodb+srv://admin:12345@cluster0.ljfwx8r.mongodb.net/Mboalab-Database?retryWrites=true&w=majority
+PORT = 5000
