+++ Main-Folder/Mboalab-Project/Backend/routes/surveyInfo.js
@@ -1,0 +1,13 @@
+const express = require("express")
+const router = express.Router();
+
+
+const {
+
+ addCheckupData
+
+} = require('../controllers/surveyInfo')
+
+router.route("/service").post(addCheckupData)
+
+module.exports = router
