+++ Main-Folder/Mboalab-Project/Backend/routes/userMessage.js
@@ -1,0 +1,12 @@
+const express = require("express")
+const router = express.Router();
+
+const {
+
+ addUserMessage
+
+} = require('../controllers/userMessage')
+
+router.route("/contact").post(addUserMessage)
+
+module.exports = router
